add temporary metadataSession to internal request params
