+++ api/common/helpers.py
@@ -8,6 +8,7 @@ from niagads.utils.list import list_to_string
 from api.common.constants import DEFAULT_PAGE_SIZE, MAX_NUM_PAGES
 from api.common.enums.cache import CacheKeyQualifier, CacheNamespace
 from api.common.enums.response_properties import ResponseContent,ResponseView, ResponseFormat
+from api.common.services.metadata_query import MetadataQueryService
 from api.common.types import Range
 
 from api.dependencies.parameters.services import InternalRequestParameters
@@ -283,5 +284,126 @@ class RouteHelper():
                 raise NotImplementedError(f'A response for view of type {str(self._responseConfig.view)} is coming soon.')
 
             
+class MetadataRouteHelper(RouteHelper):
+    """ RouteHelper extended w/Metadata queries"""
+    def __init__(self, managers: InternalRequestParameters, responseConfig: ResponseConfiguration, params: Parameters):
+        super().__init__(managers, responseConfig, params)
+        
+    async def get_track_metadata(self, rawResponse=False):
+        """ fetch track metadata; expects a list of track identifiers in the parameters"""
+        isCached = True # assuming true from the start
+        cacheKey = self._managers.cacheKey.encrypt()
+        if rawResponse:
+            cacheKey += CacheKeyQualifier.RAW
+        
+        result = await self._managers.cache.get(
+            cacheKey, namespace=self._managers.cacheKey.namespace)
+        
+        if result is None:
+            isCached = False
+        
+            tracks = self._parameters.get('_tracks',  self._parameters.get('track'))
+            tracks = tracks.split(',') if isinstance(tracks, str) else tracks
+            tracks = sorted(tracks) # best for caching & pagination
+            
+            result = await MetadataQueryService(self._managers.metadataSession) \
+                .get_track_metadata(tracks, responseType=self._responseConfig.content)
+            
+            if not rawResponse:
+                self._resultSize = len(result)
+                pageResponse = self.initialize_pagination(raiseError=False)
+                if pageResponse:
+                    sliceRange = self.slice_result_by_page()
+                    result = result[sliceRange.start:sliceRange.end]
+            
+        if rawResponse:
+            # cache the raw response
+            await self._managers.cache.set(
+                cacheKey, result, 
+                namespace=self._managers.cacheKey.namespace)
+            
+            return result
 
+        return await self.generate_response(result, isCached=isCached)
+
+
+    # FIXME: not sure if this will ever need a "rawResponse"
+    async def get_collection_track_metadata(self, rawResponse=False):
+        """ fetch track metadata for a specific collection """
+        isCached = True # assuming true from the start
+        cacheKey = self._managers.cacheKey.encrypt()
+        if rawResponse:
+            cacheKey += CacheKeyQualifier.RAW + '_' + str(rawResponse)    
+        
+        result = await self._managers.cache.get(
+            cacheKey, 
+            namespace=self._managers.cacheKey.namespace)
+        
+        if result is None:
+            isCached = False
+        
+            result = await MetadataQueryService(self._managers.metadataSession, self._managers.requestData) \
+                .get_collection_track_metadata(self._parameters.collection, self._parameters.track,
+                    responseType=self._responseConfig.content)
+            
+            if not rawResponse:
+                self._resultSize = len(result)
+                pageResponse = self.initialize_pagination(raiseError=False)
+                if pageResponse:
+                    sliceRange = self.slice_result_by_page()
+                    result = result[sliceRange.start:sliceRange.end]
+            
+        if rawResponse:
+            # cache the raw response
+            await self._managers.cache.set(
+                cacheKey, result, 
+                namespace=self._managers.cacheKey.namespace)   
+            return result
+            
+        return await self.generate_response(result, isCached=isCached)
     
+
+    async def search_track_metadata(self, rawResponse:Optional[ResponseContent] = None):
+        """ retrieve track metadata based on filter/keyword searches """
+        cacheKey = self._managers.cacheKey.encrypt()
+        content = self._responseConfig.content
+        
+        if rawResponse is not None:
+            content = rawResponse
+            cacheKey += CacheKeyQualifier.RAW + '_' + str(rawResponse)    
+        
+        result = await self._managers.cache.get(
+            cacheKey, namespace=self._managers.cacheKey.namespace)
+        
+        if result is not None:
+            return result if rawResponse else await self.generate_response(result, isCached=True) 
+                
+        offset = None
+        limit = None
+        if rawResponse is None:
+            # get counts to either return or determine pagination
+            result = await MetadataQueryService(self._managers.metadataSession) \
+                .query_track_metadata(self._parameters.assembly, 
+                    self._parameters.get('filter', None), self._parameters.get('keyword', None), ResponseContent.COUNTS)
+        
+            if content == ResponseContent.COUNTS:
+                return await self.generate_response(result, isCached=False)
+            
+            self._resultSize = result['num_tracks']
+            pageResponse = self.initialize_pagination(raiseError=False)
+            if pageResponse: # will return true if model can be paged and page is valid
+                offset = self.offset()
+                limit = self._pageSize
+            
+        result = await MetadataQueryService(self._managers.metadataSession) \
+            .query_track_metadata(self._parameters.assembly, 
+                self._parameters.get('filter', None), self._parameters.get('keyword', None), 
+                content, limit, offset)
+
+        if rawResponse is None:
+            return await self.generate_response(result, isCached=False) 
+        else: # cache the raw response before returning
+            await self._managers.cache.set(cacheKey, result, 
+                namespace=self._managers.cacheKey.namespace)
+            return result
+
