+++ api/dependencies/parameters/services.py
@@ -18,7 +18,9 @@ class InternalRequestParameters(BaseModel, arbitrary_types_allowed=True):
     request: Request
     requestData: RequestDataModel = Depends(RequestDataModel.from_request)
     session: Optional[AsyncSession] = Depends(get_none) # callable to return none, override as needed for each endpoint
+    metadataSession: Optional[AsyncSession] = Depends(get_none) # callable to return none, override as needed for each endpoint
     cacheKey: CacheKeyDataModel = Depends(CacheKeyDataModel.from_request)
     cache: Annotated[CacheManager, Depends(CACHE_MANAGER)]
+
