+++ api/routers/genomics/dependencies/parameters.py
@@ -3,13 +3,14 @@ from typing import Annotated
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.dependencies.database import DatabaseSessionManager
-
 from api.dependencies.parameters.services import InternalRequestParameters as BaseInternalRequestParameters
 
-from api.routers.genomics.common.constants import ROUTE_DATABASE
+from api.routers.genomics.common.constants import ROUTE_DATABASE, METADATA_DATABASE
 
 # initialize database and api wrapper session managers; this allows us to 
 # use connection pooling
-ROUTE_SESSION_MANAGER = DatabaseSessionManager(ROUTE_DATABASE)
+TRACK_DATA_SESSION_MANAGER = DatabaseSessionManager(ROUTE_DATABASE)
+METADATA_SESSION_MANAGER = DatabaseSessionManager(METADATA_DATABASE)
 class InternalRequestParameters(BaseInternalRequestParameters, arbitrary_types_allowed=True):
-    session: Annotated[AsyncSession, Depends(ROUTE_SESSION_MANAGER)]
+    session: Annotated[AsyncSession, Depends(TRACK_DATA_SESSION_MANAGER)]
+    metadataSession: Annotated[AsyncSession, Depends(METADATA_SESSION_MANAGER)]
